requirements.txt
opencv-python
numpy
pandas
pillow
